fix: update address - change lat, lng type float to double

See merge request s07-blockchain-contract-sub2/S07P22A706!101

+++ backend/src/main/java/com/ssafy/setak/api/request/LaundryUpdateReq.java
@@ -28,10 +28,10 @@ public class LaundryUpdateReq {
     String addrDetail;
 
     @ApiModelProperty(name = "위도")
-    float addrLat;
+    double addrLat;
 
     @ApiModelProperty(name = "경도")
-    float addrLng;
+    double addrLng;
 
     @ApiModelProperty(name = "상세설명")
     String description;
